mcp-server/requirements.txt
mcp[cli]
python-chess
